Factor out unused Errors class
